+++ app/src/main/kotlin/me/amrbashir/hijriwidget/preferences/routes/Language.kt
@@ -2,10 +2,14 @@ package me.amrbashir.hijriwidget.preferences.routes
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import me.amrbashir.hijriwidget.Preferences
 import me.amrbashir.hijriwidget.SupportedLanguage
 import me.amrbashir.hijriwidget.preferences.composables.PreferenceCategory
@@ -20,6 +24,12 @@ fun Language() {
             .fillMaxSize()
             .verticalScroll(rememberScrollState())
     ) {
+        Text(
+            "Language",
+            modifier = Modifier.padding(horizontal = 16.dp),
+            style = MaterialTheme.typography.bodySmall,
+        )
+
         for (lang in SupportedLanguage.entries) {
             PreferenceCategory(
                 label = "$lang",