+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 - Added "Calendar Calculation Method" to choose between "Islamic", "Islamic Umm al-Qura" or "Islamic Civil" methods
 - Added "Day Offset" option to adjust Hijri date by ±1 day to match local moon sightings or personal observance
+- Added "Background Color" option to adjust the widget background color
 
 ## Changed
 
+- Allow using a transparent text color when using the "Custom Color" option
 - Use `AlarmManager` instead of `WorkManager` for accurate date updates
 
 ## [v0.11.2] - 2025-03-06